更改场地金额类型

+++ gym_java/src/main/java/edu/gdou/gym_java/entity/model/Field.java
@@ -38,7 +38,7 @@ public class Field implements Serializable {
 
     @ApiModelProperty("收费标准")
     @TableField("money")
-    private Integer money;
+    private Double money;
 
     @ApiModelProperty("场地描述")
     @TableField("description")
